modified:   client/src/config.ts 	modified:   client/src/game/Game.ts 	modified:   client/src/game/entities/Unit.ts 	modified:   client/src/game/manager/Manager.ts 	modified:   client/src/pages/Battle/BattleCanvas.tsx
