+++ client/src/config.ts
@@ -21,9 +21,9 @@ export type TWINDOW = {
 const CONFIG = {
     //udsu 
     //HOST: 'http://strategy/api',
-    HOST: 'http://server/api',
+    //HOST: 'http://server/api',
     //HOST: 'http://strategy1/api',
-    // HOST: 'http://strategy/server/api',
+    HOST: 'http://strategy/server/api',
 
     CHAT_TIMESTAMP: 200, //ms
     CHAT_MAX_MESSAGE_LENGTH: 100, // максимальная длина сообщения в чате
